+++ helios/__init__.py
@@ -1,3 +1,0 @@
-"""Helios programming language toolchain."""
-
-__version__ = "0.0.1"